internal/fiat: Remove the divstep based inversion routines
A casual attempt at hand-optimizing this by inlining the 2x divstep
still doesn't get this close to the addchain based performance, and I'm
honestly satisfied with the inversion performance.

I asume this is a combination of the Go compiler's relatively shit
optimization, and fiat's Go output, but when I tried the same routines
with a Go 1.21 snapshot, performance wasn't much better.